Update Options.cs
Added switch statements to change the piece colors. 
Also added functions to display error messages if the piece color choices are the same, or if either piece color choice is the same color as the fore-color choice. 
I'll send you the file that has the images of the pieces that you'll need to make this compile.