Update Options.cs
Added to the apply button click, so that you cannot apply changes when the background/foreground are the same color, the pieces are the same color, or either piece is the same color as the foreground. Those are the last changes I was planning on making to the options class.